+++ src/Controllers/Private/Tenants.php
@@ -239,6 +239,16 @@ class Tenants extends PrivateApiController implements CrudControllerInterface
 
         $body = $this->getResourceBody($this->tenantsModel);
 
+        try {
+
+            if (!$this->user->isAdmin() && isset($body['owner']) && !$this->user->ownsTenant($params['id'])) {
+                throw new BadRequestException('Unable to update resource: Only current owner can update owner field');
+            }
+
+        } catch (UnexpectedException $e) {
+            throw new ApiServiceException($e->getMessage());
+        }
+
         if (!$this->user->isAdmin() &&
             (isset($body['domain']) || isset($body['enabled']))) {
             throw new BadRequestException('Unable to update resource: Invalid field(s)');